# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexamove Hyperlocal Delivery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  customer: "customer",
  rider: "rider",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  rider: "rider",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Shop {
  id: number;
  name: string;
  category: string;
  lat: number;
  lng: number;
  distance: number;
  eta: number;
  rating: number;
  imageUrl: string;
  isOpen: boolean;
}

export interface CreateOrderBody {
  shopId: number;
  customerLat: number;
  customerLng: number;
  items?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  preparing: "preparing",
  on_the_way: "on_the_way",
  delivered: "delivered",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  preparing: "preparing",
  on_the_way: "on_the_way",
  delivered: "delivered",
} as const;

export interface Order {
  id: number;
  customerId: number;
  shopId: number;
  /** @nullable */
  riderId?: number | null;
  status: OrderStatus;
  customerLat: number;
  customerLng: number;
  shopName: string;
  /** @nullable */
  riderName?: string | null;
  createdAt: string;
}

export interface UpdateLocationBody {
  lat: number;
  lng: number;
  /** @nullable */
  orderId?: number | null;
}

export interface Rider {
  id: number;
  name: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  isAvailable: boolean;
}

export type ListShopsParams = {
  lat: number;
  lng: number;
  radius?: number;
};

export type ListAvailableRidersParams = {
  lat: number;
  lng: number;
};