# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexamove Hyperlocal Delivery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: shops
    description: Shop operations
  - name: orders
    description: Order operations
  - name: riders
    description: Rider operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List nearby shops
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          required: false
          schema:
            type: number
            default: 3
      responses:
        "200":
          description: List of nearby shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"

  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (rider only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /riders/available:
    get:
      operationId: listAvailableRiders
      tags: [riders]
      summary: List available riders near a location
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Available riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"

  /riders/location:
    post:
      operationId: updateRiderLocation
      tags: [riders]
      summary: Update rider's current GPS location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rider"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, rider]
      required:
        - name
        - email
        - password
        - role

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, rider]
      required:
        - id
        - name
        - email
        - role

    Shop:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        lat:
          type: number
        lng:
          type: number
        distance:
          type: number
        eta:
          type: integer
        rating:
          type: number
        imageUrl:
          type: string
        isOpen:
          type: boolean
      required:
        - id
        - name
        - category
        - lat
        - lng
        - distance
        - eta
        - rating
        - imageUrl
        - isOpen

    CreateOrderBody:
      type: object
      properties:
        shopId:
          type: integer
        customerLat:
          type: number
        customerLng:
          type: number
        items:
          type: string
      required:
        - shopId
        - customerLat
        - customerLng

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [preparing, on_the_way, delivered]
      required:
        - status

    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        shopId:
          type: integer
        riderId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [preparing, on_the_way, delivered]
        customerLat:
          type: number
        customerLng:
          type: number
        shopName:
          type: string
        riderName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerId
        - shopId
        - status
        - customerLat
        - customerLng
        - shopName
        - createdAt

    UpdateLocationBody:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        orderId:
          type: ["integer", "null"]
      required:
        - lat
        - lng

    Rider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isAvailable:
          type: boolean
      required:
        - id
        - name
        - isAvailable