# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexamove Hyperlocal Delivery API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterUserBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  role: zod.enum(["customer", "rider"]),
});

/**
 * @summary Login with email and password
 */
export const LoginUserBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginUserResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    role: zod.enum(["customer", "rider"]),
  }),
});

/**
 * @summary Get current user info
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["customer", "rider"]),
});

/**
 * @summary List nearby shops
 */
export const listShopsQueryRadiusDefault = 3;

export const ListShopsQueryParams = zod.object({
  lat: zod.coerce.number(),
  lng: zod.coerce.number(),
  radius: zod.coerce.number().default(listShopsQueryRadiusDefault),
});

export const ListShopsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  distance: zod.number(),
  eta: zod.number(),
  rating: zod.number(),
  imageUrl: zod.string(),
  isOpen: zod.boolean(),
});
export const ListShopsResponse = zod.array(ListShopsResponseItem);

/**
 * @summary Get a shop by ID
 */
export const GetShopParams = zod.object({
  id: zod.coerce.number(),
});

export const GetShopResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  lat: zod.number(),
  lng: zod.number(),
  distance: zod.number(),
  eta: zod.number(),
  rating: zod.number(),
  imageUrl: zod.string(),
  isOpen: zod.boolean(),
});

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  shopId: zod.number(),
  customerLat: zod.number(),
  customerLng: zod.number(),
  items: zod.string().optional(),
});

/**
 * @summary List current user's orders
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  shopId: zod.number(),
  riderId: zod.number().nullish(),
  status: zod.enum(["preparing", "on_the_way", "delivered"]),
  customerLat: zod.number(),
  customerLng: zod.number(),
  shopName: zod.string(),
  riderName: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Get an order by ID
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  shopId: zod.number(),
  riderId: zod.number().nullish(),
  status: zod.enum(["preparing", "on_the_way", "delivered"]),
  customerLat: zod.number(),
  customerLng: zod.number(),
  shopName: zod.string(),
  riderName: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update order status (rider only)
 */
export const UpdateOrderStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateOrderStatusBody = zod.object({
  status: zod.enum(["preparing", "on_the_way", "delivered"]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  shopId: zod.number(),
  riderId: zod.number().nullish(),
  status: zod.enum(["preparing", "on_the_way", "delivered"]),
  customerLat: zod.number(),
  customerLng: zod.number(),
  shopName: zod.string(),
  riderName: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary List available riders near a location
 */
export const ListAvailableRidersQueryParams = zod.object({
  lat: zod.coerce.number(),
  lng: zod.coerce.number(),
});

export const ListAvailableRidersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  isAvailable: zod.boolean(),
});
export const ListAvailableRidersResponse = zod.array(
  ListAvailableRidersResponseItem,
);

/**
 * @summary Update rider's current GPS location
 */
export const UpdateRiderLocationBody = zod.object({
  lat: zod.number(),
  lng: zod.number(),
  orderId: zod.number().nullish(),
});

export const UpdateRiderLocationResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  lat: zod.number().nullish(),
  lng: zod.number().nullish(),
  isAvailable: zod.boolean(),
});